+++ Conductor/Autoload.php
@@ -6,7 +6,7 @@
  * phpab --var namespace=Conductor --template Conductor/Autoload.php.in --output Conductor/Autoload.php Conductor
  * 
  * @copyright 2012 Clay Loveless <clay@php.net>
- * @license   http://opensource.org/licenses/MIT MIT License
+ * @license   http://claylo.mit-license.org/2012/ MIT License
  */
 namespace Conductor;